Use Alien::SWIProlog for configuration
This removes the old configuration code in favour of Alien::SWIProlog.
One reason is that this is more correct for newer versions of SWI-Prolog
under different configurations. Since older versions of SWI-Prolog (pre
version 7) are no longer supported, the older configuration code is no
longer useful.

Another reason is that by using this, it will be easy to use SWI-Prolog
with non-threaded Perls.